Added -R option which provides package reviewing
The package review shows first the .info and README in the sbopkg
respository. Then the package file (dep file) is displayed. Both sets of files
are displayed with a pager so they can be reviewed. After the pager is exited,
the user is prompted for adding the package to the REVIEWED list.