Fixed a number of issues with the menu-driven system.
The display_dep_menu() function now allows for options to be
disabled. This is needed for cases where a package DB list is being
iterated over, and we cannot remove a dep file or it breaks things
(and doesn't make sense given the situation). Other situations may
arise as well.

The msg_string module was added for default display_dep_menu()
messages.