Fixed issue with packages checking potentially removed packages during removal.
This needs to be updated, to see if we are able to remove any of the
parent packages during removal. We are currently ignoring those,
parent packages in the dependency tree, but really, the parents should
be checked first if they can be removed, and then added to a removal
list, where from the removal list, we can see if it can be removed.