Implementation of the base functionality for the rework.
We now have a graph implementation for the package dependencies. The
package DB has been updated to include all packages in SBo repo along
with a checksum of the README and REQUIRES entry from the info
file. The reviewed DB contains the checksum from the last time the
package was reviewed, so that it can be checked against any SBo
updates and reviewed again. This requires that the package DB is
updated after SBo updates.

Dependency files are automatically created if they do not exist. Every
package in SBo will now have a dependency file.