Added graph processing for package management.
The graph consists of all packages in the SBo repo as nodes in the
graph. Connections are made as needed based on the processing. The
PKGDB will be used to load the graph, which includes a CRC of the
README and info files. The CRC can be used to check for changes in the
case of updates to the SBo repo.

All package processing is currently done by search for nodes in the
graph data structure. This does not handle meta packages since they
are represented only in the DEPDIR. To handle meta packages the graph
data structure will need to include the SBo packages along with meta
packages; they must be kept separately though, since if we output
PKGDB, they will be include erroneously if they are in the same node
set.

A search for a package will require the following steps.

  1) search the SBo packages; if not present goto 2); otherwise done.
  2) search DEPDIR for the dep file; if present goto 3); if not
     present handle missing package
  3) check for METAPKG in the dep file; if present, then goto 4);
     otherwise, handle missing package
  4) create new meta package node and add to the graph; done