Updated the packages to be stored in a graph represented by two vectors (sbo pkgs and meta pkgs).
The graph processing now separates the sbo pkgs from the meta
pkgs. They are handled seperately since they differ in how they are
specified.

The pkg node iterator is used to iterate through the graph for
required or parent package nodes. The distance evaluation currently
includes the meta packages, but this is not necessarily what is
desired.