Updated the DFS graph iterator to correctly support the forward and reverse visitiation ordering.
The iterator returns an edge node (edge_node) of the currently visited
node (cur_node) which provides the digraph of

     cur_node -> edge_node

The cur_node will always be marked GREY and the edge_node will be GREY
or BLACK for reverse ordering or WHITE or BLACK for forward
ordering. A BLACK edge node will only ever be returned if
PKG_ITER_REQ_NEAREST is set.